style: add the background stains
