+++ src/components/Campaign/Hero/styles.module.css
@@ -1,3 +1,20 @@
+.gridContainer {
+  justify-content: space-between;
+  gap: 80px;
+  position: relative;
+}
+
+.spot {
+  position: absolute;
+  right: -35px;
+  top: -110px;
+  z-index: -1;
+  width: 800px;
+  height: 800px;
+  background-image: radial-gradient(rgba(18, 255, 128, 0.3), rgba(18, 19, 18, 0.15) 60%, rgba(18, 19, 18, 1) 100%);
+  filter: blur(30px);
+}
+
 .textBlock {
   display: flex;
   gap: 24px;
