+++ src/components/Campaign/LogoTextBlockCentered/styles.module.css
@@ -3,6 +3,17 @@
   flex-direction: column;
   align-items: center;
   gap: 24px;
+  position: relative;
+}
+
+.spot {
+  position: absolute;
+  top: -130px;
+  z-index: -1;
+  width: 400px;
+  height: 400px;
+  background-image: radial-gradient(rgba(18, 255, 128, 0.3), rgba(18, 19, 18, 0.15) 60%, rgba(18, 19, 18, 1) 100%);
+  filter: blur(30px);
 }
 
 .cta {
