+++ src/components/Campaign/RoundCardGrid/styles.module.css
@@ -1,3 +1,19 @@
+.gridContainer {
+  margin-top: 32px;
+  position: relative;
+}
+
+.spot {
+  position: absolute;
+  left: -200px;
+  top: -450px;
+  z-index: -1;
+  width: 800px;
+  height: 800px;
+  background-image: radial-gradient(rgba(18, 255, 128, 0.3), rgba(18, 19, 18, 1) 80%);
+  filter: blur(50px);
+}
+
 .card {
   border: 1px solid var(--mui-palette-border-main);
   border-radius: 24px;
