+++ src/components/Campaign/TitleTweets/styles.module.css
@@ -1,15 +1,29 @@
+.titleWrapper {
+  position: relative;
+  margin: 0 auto;
+  width: fit-content;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 60px;
+}
+
+.spot {
+  position: absolute;
+  top: -200px;
+  z-index: -1;
+  width: 700px;
+  height: 700px;
+  background-image: radial-gradient(rgba(18, 255, 128, 0.3), rgba(18, 19, 18, 1) 100%);
+  filter: blur(50px);
+}
+
 .cta {
   margin-bottom: 40px;
-  text-align: center;
   background: linear-gradient(260.13deg, #12ff80 1.24%, #5fddff 102.14%);
   background-clip: text;
   -webkit-background-clip: text;
   color: transparent;
-}
-
-.title {
-  text-align: center;
-  margin-bottom: 60px;
 }
 
 .tweetsGrid {
